AND gate using 2 NAND gates.
So the logic behind this chip is really simple. Since this is the first chip, we don't have a choice but to reuse the NAND chip. Of course how many times we use it matters too. If you look at the truth table of NAND carefully then we can see that it's just an AND gate followed by a NOT gate so we just have to reverse the process. The last low of the table is usually as the follows: 1 NAND 1 == 0 which is very essential because we can just take the output of 1 NAND gate and NAND it by itself. This will immediately satisfy the first 3 rows of an AND Gate (from AND's Truth Table) and consequently will do for the last row too.

This will be the case for other chips too i.e.; we will be following a same methodology. A bit of logic, finding patterns in the truth tables and lastly, trying out different combinations- hoping that at least one would work.