NOT gate using 1 NAND gate.
This is pretty simple and one can deduce NOT gate just by looking at the table for NAND. 0 NAND 0 = 1 and 1 NAND 1 = 0. We can see clearly that we're getting the opposite of each input (whether that be 0 or 1) so we can just NAND the single input, the result will automatically will have the opposite which encapsulates the true essence of a NOT gate.