OR gate from 3 NAND gates.
The logic behind this chip is really simple. We NOT the inputs and just NAND them. Here I am NOT'ing the inputs using 2 NAND gates (first two) but you can just use NOT gates too ( given that you've build them prior to this chip). After we get the 2 results, we can NAND them using a third NAND chip and that will give us the desired result. 

If we look at the truth table of NAND, the middle 2 results ( 1 NAND 0  = 1 and 0 NAND 1 = 1) are same as OR but we get trouble when dealing with first and last ( 0 NAND 0  = 1 and 1 NAND 1 = 0). This problem is solved by inversing the both inputs as discussed above.