XOR chip made from 3 different implementations.
This chip is quite interesting, Here I have presented 3 implementations. Before going over the logic of my implementation (how I understand it), I would like to give them due credit. 

Credit of 2nd implementation: https://people.duke.edu/~nts9/logicgates/Xor.hdl

Credit of 3rd implementation: https://github.com/jcoglan/nand2tetris/blob/master/projects/01/Xor.hdl

For my approach, If we look at the XOR truth table then last 3 rows are similar to NAND which is a good news. When I understood this, I was determined to use only the NAND gates. First 3 of the 4 NAND gates are basically inversing the different input (0 1) but it is also inversing the 0 0 input to 1 1 and not changing the 1 1 input. If we get different outputs, we will surely get a 1 from the 4th NAND chip. If we get similar outputs which would already be in 1 1 (even if the starting inputs were 0 0, it would be converted to 1 1 by the first 3 NAND chips as described earlier), we can just NAND them and the result would be 0 which truly encapsulates the essence of XOR gate i.e.; 0 on same inputs and 1 on different inputs.