MUX taking 2-bit selector and 8 inputs.
The logic of this and the next chip (Mux8Way16) is very similar. First map out all the inputs in 2 rows so A B would be in first row and C D would be in second row. (A, B, C, and D are inputs). In the first 2 MUX, we're just selecting each element from each row then based on the least significant bit, we're selecting, which row's selected bit to select. Least significant bit is the left most bit if we're moving from right to left. 

If you think about it, it's amazing. If the LSB is 0, we only can take inputs from first row and if its 1, we only deal with the 2nd row. This is essentially what we're doing here. We're first selecting 2 inputs from both of the rows, one of these inputs will surely be the desired output. Why are we not sure? Because we don't know which row to take the selected bit from and this is basically where LSB comes into play. We mux that too using the LSB and we got out desired output where it should be.