DMUX using 1 NOT gate and 2 AND gates.
I took help from this video: https://www.youtube.com/watch?v=t3Ed13z9uz8. It is just like how we made MUX but we don't do OR at the end but leave the AND gates as they are since we need 2 outputs. 

Here the guy used K-Mapping again, seems like it is pretty useful technique to make chips.