FULL ADDER USING 3 NOT, 13 AND and 5 OR gates.
To do this, we have to go back to project 1 lectures where boolean algebra was discussed but I found a better way to form the chips of a specific function (which is basically the outputs for sum and carry). That trick is K-mapping. You can watch this tutorial, this has everything you need, after designing the basic logic expression, you can dry run it and finally implement it in the HDL program.

Link to K-map tutorial: https://www.youtube.com/watch?v=RO5alU6PpSU