ADDing 2 16-bit busses using Full and Half Adders.
The chip is really easy to make, the only problem we face is with the first bit (least significant bit). For LSB, we can't use a FullAdder because we only have 2 inputs and a FullAdder requires 3 inputs. How can we solve this? Just use a HalfAdder for the first bit and after that, send the carry to next FullAdder while also sending 2 inputs from the 2 busses. 

Initially my code didn't work, even though the logic was completely same. Perhaps it was because I was attaching the carry bit to "a" pin of FullAdders and not "c" (like in this code). Nonetheless, the code is from: https://people.duke.edu/~nts9/logicgates/Add16.hdl